update paths for setup scripts
