+++ readme.md
@@ -54,7 +54,7 @@ To get a local copy up and running follow these simple example steps.
 
 1. Seed the database! (you may need to change permissions on your local machine)
    ```sh
-   ./db/db_seed.sh
+   ./bin/db_seed.sh
    ```
 
 ---
@@ -78,8 +78,7 @@ To get a local copy up and running follow these simple example steps.
 1. In the root directory of the app folder, create a .env and set values for CHECKING_PORT and VALIDATION_PORT
 
    ```
-   echo "CHECKING_PORT=5000" >> .env && \
-   echo "VALIDATION_PORT=5500" >> .env
+   echo -e "CHECKING_PORT=5000\nVALIDATION_PORT=5500" > .env
    ```
 
 2. Run the services (or both!)
